# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: 中建智慧安全平台 MVP
  version: 0.1.0
paths:
...
  /api/v1/profile/project/{project_id}:
    get:
      tags:
      - profiles
      summary: Project Profile
      operationId: project_profile_api_v1_profile_project__project_id__get
      parameters:
      - name: project_id
        in: path
        required: true
        schema:
          type: string
          title: Project Id
      - name: X-Mock-User-Id
        in: header
        required: false
        schema:
          type: string
          default: mock-admin
          title: X-Mock-User-Id
      - name: X-Mock-User-Name
        in: header
        required: false
        schema:
          type: string
          default: 平台管理员
          title: X-Mock-User-Name
      - name: X-Tenant-Id
        in: header
        required: false
        schema:
          type: string
          default: CSCEC
          title: X-Tenant-Id
      - name: X-Org-Path
        in: header
        required: false
        schema:
          type: string
          default: CSCEC
          title: X-Org-Path
      - name: X-Role
        in: header
        required: false
        schema:
          type: string
          default: platform_admin
          title: X-Role
      - name: X-Data-Scope
        in: header
        required: false
        schema:
          type: string
          default: tenant
          title: X-Data-Scope
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/profile/ranking/projects:
    get:
      tags:
      - profiles
      summary: Project Ranking
      operationId: project_ranking_api_v1_profile_ranking_projects_get
      parameters:
      - name: X-Mock-User-Id
        in: header
        required: false
        schema:
          type: string
          default: mock-admin
          title: X-Mock-User-Id
      - name: X-Mock-User-Name
        in: header
        required: false
        schema:
          type: string
          default: 平台管理员
          title: X-Mock-User-Name
      - name: X-Tenant-Id
        in: header
        required: false
        schema:
          type: string
          default: CSCEC
          title: X-Tenant-Id
      - name: X-Org-Path
        in: header
        required: false
        schema:
          type: string
          default: CSCEC
          title: X-Org-Path
      - name: X-Role
        in: header
        required: false
        schema:
          type: string
          default: platform_admin
          title: X-Role
      - name: X-Data-Scope
        in: header
        required: false
        schema:
          type: string
          default: tenant
          title: X-Data-Scope
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/memory/session:
    post:
      tags:
      - memory
      summary: Upsert Memory Session
      operationId: upsert_memory_session_api_v1_memory_session_post
      parameters:
      - name: X-Mock-User-Id
        in: header
        required: false
        schema:
          type: string
          default: mock-admin
          title: X-Mock-User-Id
      - name: X-Mock-User-Name
        in: header
        required: false
        schema:
          type: string
          default: 平台管理员
          title: X-Mock-User-Name
      - name: X-Tenant-Id
        in: header
        required: false
        schema:
          type: string
          default: CSCEC
          title: X-Tenant-Id
      - name: X-Org-Path
        in: header
        required: false
        schema:
          type: string
          default: CSCEC
          title: X-Org-Path
      - name: X-Role
        in: header
        required: false
        schema:
          type: string
          default: platform_admin
          title: X-Role
      - name: X-Data-Scope
        in: header
        required: false
        schema:
          type: string
          default: tenant
          title: X-Data-Scope
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MemorySessionUpsertRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Upsert Memory Session Api V1 Memory Session Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    get:
      tags:
      - memory
      summary: Read Memory Session
      operationId: read_memory_session_api_v1_memory_session_get
      parameters:
      - name: session_id
        in: query
        required: true
        schema:
          type: string
          minLength: 1
          maxLength: 64
          title: Session Id
      - name: X-Mock-User-Id
        in: header
        required: false
        schema:
          type: string
          default: mock-admin
          title: X-Mock-User-Id
      - name: X-Mock-User-Name
        in: header
        required: false
        schema:
          type: string
          default: 平台管理员
          title: X-Mock-User-Name
      - name: X-Tenant-Id
        in: header
        required: false
        schema:
          type: string
          default: CSCEC
          title: X-Tenant-Id
      - name: X-Org-Path
        in: header
        required: false
        schema:
          type: string
          default: CSCEC
          title: X-Org-Path
      - name: X-Role
        in: header
        required: false
        schema:
          type: string
          default: platform_admin
          title: X-Role
      - name: X-Data-Scope
        in: header
        required: false
        schema:
          type: string
          default: tenant
          title: X-Data-Scope
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Read Memory Session Api V1 Memory Session Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    MemoryMessage:
      properties:
        role:
          type: string
          enum:
          - system
          - user
          - assistant
          - tool
          title: Role
        content:
          type: string
          minLength: 1
          title: Content
        metadata:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Metadata
      type: object
      required:
      - role
      - content
      title: MemoryMessage
    MemorySessionUpsertRequest:
      properties:
        session_id:
          type: string
          maxLength: 64
          minLength: 1
          title: Session Id
        message:
          $ref: '#/components/schemas/MemoryMessage'
        context:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Context
        summary:
          anyOf:
          - type: string
          - type: 'null'
          title: Summary
      type: object
      required:
      - session_id
      - message
      title: MemorySessionUpsertRequest
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...